split SpinalSimTester out from misc tester.
