+++ tester/src/test/scala/spinal/tester/scalatest/SpinalSimMiscTester.scala
@@ -6,7 +6,7 @@ import spinal.core._
 import spinal.sim._
 import spinal.core.sim.{SpinalSimConfig, _}
 import spinal.lib.{BufferCC, OHMasking, SetFromFirstOne}
-import spinal.tester
+import spinal.tester.SpinalAnyFunSuite
 import spinal.tester.scalatest
 
 import scala.concurrent.{Await, Future}
@@ -28,47 +28,6 @@ object SpinalSimMiscTester{
 
 }
 
-abstract class SpinalSimTester{
-  def SimConfig : SpinalSimConfig
-  def durationFactor : Double
-  def designFactor : Double
-  def prefix : String
-  def language : SpinalMode
-}
-
-object SpinalSimTesterGhdl extends SpinalSimTester{
-  override def SimConfig: SpinalSimConfig = spinal.core.sim.SimConfig.withGhdl
-  override def durationFactor: Double = 0.005
-  override def designFactor: Double = 0.05
-  override def prefix: String = "ghdl_"
-  override def language: SpinalMode = VHDL
-}
-
-object SpinalSimTesterIVerilog extends SpinalSimTester{
-  override def SimConfig: SpinalSimConfig = spinal.core.sim.SimConfig.withIVerilog
-  override def durationFactor: Double = 0.005
-  override def designFactor: Double = 0.05
-  override def prefix: String = "iverilog_"
-  override def language: SpinalMode = Verilog
-}
-
-object SpinalSimTesterVerilator extends SpinalSimTester{
-  override def SimConfig: SpinalSimConfig = spinal.core.sim.SimConfig.withVerilator
-  override def durationFactor: Double = 0.5
-  override def designFactor: Double = 0.5
-  override def prefix: String = "verilator_"
-  override def language: SpinalMode = Verilog
-}
-
-object SpinalSimTester{
-
-  def apply(body :  => SpinalSimTester => Unit): Unit = {
-    body(SpinalSimTesterGhdl)
-    body(SpinalSimTesterIVerilog)
-    body(SpinalSimTesterVerilator)
-  }
-}
-
 class SpinalSimTesterTest extends SpinalAnyFunSuite {
   SpinalSimTester{ env =>
     import env._
@@ -79,45 +38,11 @@ class SpinalSimTesterTest extends SpinalAnyFunSuite {
   }
 }
 
-class SpinalSimFunSuite extends SpinalAnyFunSuite{
-  var tester : SpinalSimTester = null
-  def SimConfig = tester.SimConfig
-  var durationFactor = 0.0
-  var ghdlEnabled = true
-  var iverilogEnabled = true
-
-  def onlyVerilator(): Unit ={
-    iverilogEnabled = false
-    ghdlEnabled = false
-  }
-
-  def withAll(): Unit ={
-    iverilogEnabled = true
-    ghdlEnabled = true
-  }
-  def test(testName: String)(testFun: => Unit): Unit = {
-    super.test("verilator_" + testName) {
-      tester = SpinalSimTesterVerilator
-      durationFactor = SpinalSimTesterVerilator.durationFactor
-      testFun
-    }
-    if(ghdlEnabled) super.test("ghdl_" + testName) {
-      tester = SpinalSimTesterGhdl
-      durationFactor = SpinalSimTesterGhdl.durationFactor
-      testFun
-    }
-    if(iverilogEnabled) super.test("iverilog_" + testName) {
-      tester = SpinalSimTesterIVerilog
-      durationFactor = SpinalSimTesterIVerilog.durationFactor
-      testFun
-    }
-  }
-}
 
 class SpinalSimMiscTester extends SpinalAnyFunSuite {
   SpinalSimTester { env =>
     import env._
-    var compiled: SimCompiled[tester.scalatest.SpinalSimMiscTester.SpinalSimMiscTesterCounter] = null
+    var compiled: SimCompiled[SpinalSimMiscTester.SpinalSimMiscTesterCounter] = null
 
     test(prefix + "testForkSensitive") {
       SimConfig.compile(new Component {
@@ -142,7 +67,7 @@ class SpinalSimMiscTester extends SpinalAnyFunSuite {
 
 
     test(prefix + "compile") {
-      compiled = SimConfig.compile(new tester.scalatest.SpinalSimMiscTester.SpinalSimMiscTesterCounter)
+      compiled = SimConfig.compile(new SpinalSimMiscTester.SpinalSimMiscTesterCounter)
     }
 
     def doStdtest(name: String): Unit = {
@@ -297,7 +222,7 @@ class SpinalSimMiscTester extends SpinalAnyFunSuite {
     }
 
     test(prefix + "testRecompile1") {
-      SimConfig.doSim(new tester.scalatest.SpinalSimMiscTester.SpinalSimMiscTesterCounter)(dut => {
+      SimConfig.doSim(new SpinalSimMiscTester.SpinalSimMiscTesterCounter)(dut => {
         dut.clockDomain.forkStimulus(10)
 
         var counterModel = 0
@@ -316,7 +241,7 @@ class SpinalSimMiscTester extends SpinalAnyFunSuite {
 
 
     test(prefix + "testRecompile2") {
-      SimConfig.doSim(new tester.scalatest.SpinalSimMiscTester.SpinalSimMiscTesterCounter)(dut => {
+      SimConfig.doSim(new SpinalSimMiscTester.SpinalSimMiscTesterCounter)(dut => {
         dut.clockDomain.forkStimulus(10)
 
         var counterModel = 0
